Refactor/Corrected public api. Merged products and categories api into one file
