--- js/shared/publicApi/caterogies.api.js
@@ -1,7 +1,0 @@
-const baseUrl = 'https://fakestoreapi.com'
-
-const categoriesAPI = {
-    allCategories: `${baseUrl}/products/categories`
-}
-
-export default categoriesAPI
