+++ js/shared/publicApi/index.js
@@ -1,5 +1,13 @@
-import productsAPI from "./products.api.js";
-import categoriesAPI from "./caterogies.api.js";
+const baseUrl = 'https://fakestoreapi.com'
+
+const productsAPI = {
+    allProducts: `${baseUrl}/products`,
+}
+
+const categoriesAPI = {
+    allCategories: `${baseUrl}/products/categories`
+}
+
 
 export {
     productsAPI,
