--- js/shared/publicApi/products.api.js
@@ -1,7 +1,0 @@
-const baseUrl = 'https://fakestoreapi.com'
-
-const productsAPI = {
-    allProducts: `${baseUrl}/products`,
-}
-
-export default productsAPI
